test: split CLI workspace test support
